move metric to internal
